pyproject.toml
[tool.poetry]
name = "dreamdeck"
version = "0.1.0"
description = "AI Storytelling done right"
authors = ["Danny Key <danny@dannydabbles.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "src", from = "." },
    { include = "src.*", from = "." }
]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.14.3"
chromadb = "^0.4.24"
sentence-transformers = "^3.3.1"
pydantic = "2.10.1"
langchain-core = "0.3.25"
aiohttp = "^3.11.10"
sqlalchemy = "^2.0.36"
langchain-community = "^0.3.12"
langgraph = "^0.2.59"
langchain-openai = "^0.2.12"
langchain-huggingface = "^0.1.2"
langchain-chroma = "^0.1.4"
aiosqlite = "^0.20.0"
asyncpg = "^0.30.0"
azure-storage-blob = "^12.24.0"
boto3 = "^1.35.99"
chainlit = "^2.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
